Move polygon types and utils to map planning util directory.
