+++ frontend/src/features/map_planning/utils/PolygonTypes.ts
